+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 0.1.1
  - setters for pattern name, track & line count
+ - getter for current line
 
 ## 0.1.0
  - update to Sunvox lib release v2.1b